backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch>=2.2.0
torchvision>=0.17.0
safetensors>=0.4.1
Pillow>=10.2.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
python-multipart>=0.0.6
opencv-python>=4.9.0
